emg/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
distlib==0.3.8
Django==5.0.6
django-cors-headers==4.4.0
django-oauth-toolkit==2.4.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
filelock==3.15.4
idna==3.7
inflection==0.5.1
jwcrypto==1.5.6
oauthlib==3.2.2
packaging==24.1
platformdirs==4.2.2
pycparser==2.22
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.7
requests==2.32.3
setuptools==70.2.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
virtualenv==20.26.3
